+++ monolithic/monolithic-layered/src/main/java/com/epam/training/microservices/monolithic/web/crud/single/ViewSingleSupport.java
@@ -2,12 +2,8 @@ package com.epam.training.microservices.monolithic.web.crud.single;
 
 import com.epam.training.microservices.monolithic.exception.ResourceNotFoundException;
 import com.epam.training.microservices.monolithic.web.crud.CrudSupport;
-import java.util.Optional;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 public interface ViewSingleSupport<T> extends CrudSupport<T> {
@@ -26,18 +22,5 @@ public interface ViewSingleSupport<T> extends CrudSupport<T> {
     modelAndView.addObject("submitTarget", getSubmitTarget());
     modelAndView.addObject("templateParams", templateParams);
     return modelAndView;
-  }
-
-  @PostMapping("")
-  default String saveSingle(@ModelAttribute T item) {
-    getService().save(item);
-    return "redirect:" + getSubmitTarget();
-  }
-
-  default String getSubmitTarget() {
-    return Optional.ofNullable(this.getClass().getDeclaredAnnotation(RequestMapping.class))
-        .map(RequestMapping::value)
-        .map(values -> values[0])
-        .orElseThrow(() -> new RuntimeException("No RequestMapping annotation"));
   }
 }
