+++ monolithic/monolithic-layered/src/main/java/com/epam/training/microservices/monolithic/web/crud/single/details/ViewSingleDetailsSupport.java
@@ -4,10 +4,9 @@ import com.epam.training.microservices.monolithic.exception.ResourceNotFoundExce
 import com.epam.training.microservices.monolithic.web.crud.CrudSupport;
 import com.epam.training.microservices.monolithic.web.crud.single.ViewSingleTemplateParams;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.servlet.ModelAndView;
-
-import java.util.Optional;
 
 public interface ViewSingleDetailsSupport<T> extends CrudSupport<T> {
   ViewSingleDetailsTemplateParams getViewSingleTemplateParams(T parent);
@@ -26,18 +25,5 @@ public interface ViewSingleDetailsSupport<T> extends CrudSupport<T> {
     modelAndView.addObject("submitTarget", getSubmitTarget());
     modelAndView.addObject("templateParams", templateParams);
     return modelAndView;
-  }
-
-  @PostMapping("")
-  default String saveSingle(@ModelAttribute T item) {
-    getService().save(item);
-    return "redirect:" + getSubmitTarget();
-  }
-
-  default String getSubmitTarget() {
-    return Optional.ofNullable(this.getClass().getDeclaredAnnotation(RequestMapping.class))
-        .map(RequestMapping::value)
-        .map(values -> values[0])
-        .orElseThrow(() -> new RuntimeException("No RequestMapping annotation"));
   }
 }
